拆分 DynamicWidgetBuilder，各部分各司其职
