+++ fair/lib/fair.dart
@@ -22,4 +22,8 @@ export 'src/runtime/fair_runtime_impl.dart';
 export 'src/runtime/plugin/fair_plugin_suggar.dart';
 export 'src/runtime/plugin/fair_plugin_m.dart';
 export 'src/render/domain.dart';
-export 'src/render/builder.dart';
+export 'src/render/builder/builder.dart';
+export 'src/render/builder/common.dart';
+export 'src/render/builder/function.dart';
+export 'src/render/builder/widget.dart';
+export 'src/type.dart';
