+++ fair/lib/src/render/builder/builder.dart
@@ -1,0 +1,25 @@
+/*
+ * Copyright (C) 2005-present, 58.com.  All rights reserved.
+ * Use of this source code is governed by a BSD type license that can be
+ * found in the LICENSE file.
+ */
+import 'package:fair/src/internal/bind_data.dart';
+import 'package:fair/src/render/domain.dart';
+import 'package:fair/src/render/proxy.dart';
+import 'package:flutter/material.dart';
+
+abstract class Binding {}
+
+abstract class DynamicBuilder {
+  final String? tag;
+  final ProxyMirror? proxyMirror;
+  final String? page;
+  final BindingData? bound;
+  final String? bundle;
+
+  DynamicBuilder(this.tag, this.proxyMirror, this.page, this.bound,
+      {this.bundle});
+
+  dynamic convert(BuildContext context, Map map,  Map? methodMap,
+      {Domain? domain});
+}
